+++ base2d_kinematics/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(kinematics_2d)
+project(base2d_kinematics)
 
 # Default to C99
 if(NOT CMAKE_C_STANDARD)
@@ -19,17 +19,17 @@ set(CMAKE_CXX_FLAGS " ${CMAKE_CXX_FLAGS_INIT} -fPIC")
 
 find_package(ament_cmake REQUIRED)
 find_package(geometry_msgs REQUIRED)
-find_package(kinematics_2d_msgs REQUIRED)
+find_package(base2d_kinematics_msgs REQUIRED)
 find_package(nav_2d_msgs REQUIRED)
 find_package(rclcpp REQUIRED)
 
-add_library(kinematics_2d src/kinematic_parameters.cpp src/noise_model.cpp)
-target_include_directories(kinematics_2d PUBLIC
+add_library(base2d_kinematics src/kinematic_parameters.cpp src/noise_model.cpp)
+target_include_directories(base2d_kinematics PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>
 )
-ament_target_dependencies(kinematics_2d
-  kinematics_2d_msgs
+ament_target_dependencies(base2d_kinematics
+  base2d_kinematics_msgs
   rclcpp
   geometry_msgs
   nav_2d_msgs
@@ -40,9 +40,9 @@ target_include_directories(kinematics_publisher PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>
 )
-target_link_libraries(kinematics_publisher kinematics_2d)
+target_link_libraries(kinematics_publisher base2d_kinematics)
 ament_target_dependencies(kinematics_publisher
-  kinematics_2d_msgs
+  base2d_kinematics_msgs
   rclcpp
   geometry_msgs
   nav_2d_msgs
@@ -51,9 +51,9 @@ install(TARGETS kinematics_publisher
         DESTINATION lib/${PROJECT_NAME}
 )
 
-ament_export_targets(export_kinematics_2d HAS_LIBRARY_TARGET)
-install(TARGETS kinematics_2d
-        EXPORT export_kinematics_2d
+ament_export_targets(export_base2d_kinematics HAS_LIBRARY_TARGET)
+install(TARGETS base2d_kinematics
+        EXPORT export_base2d_kinematics
         ARCHIVE DESTINATION lib
         LIBRARY DESTINATION lib
         RUNTIME DESTINATION bin
@@ -62,6 +62,6 @@ install(DIRECTORY include/
         DESTINATION include
 )
 ament_export_include_directories(include)
-ament_export_dependencies(kinematics_2d_msgs)
+ament_export_dependencies(base2d_kinematics_msgs)
 
 ament_package()
