+++ base2d_kinematics/README.md
@@ -1,3 +1,3 @@
-# kinematics_2d
+# base2d_kinematics
 
 Based in part on [`dwb_plugins::KinematicParameters`](https://github.com/locusrobotics/robot_navigation/tree/noetic/dwb_plugins)
