+++ base2d_kinematics/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kinematics_2d</name>
+  <name>base2d_kinematics</name>
   <version>0.1.0</version>
   <description>Implementation of simple kinematics in two and half dimensions</description>
   <maintainer email="davidvlu@gmail.com">David V. Lu!!</maintainer>
@@ -9,7 +9,7 @@
 
   <buildtool_depend>ament_cmake</buildtool_depend>
   <depend>geometry_msgs</depend>
-  <depend>kinematics_2d_msgs</depend>
+  <depend>base2d_kinematics_msgs</depend>
   <depend>nav_2d_msgs</depend>
   <depend>rclcpp</depend>
