+++ base2d_kinematics_msgs/CHANGELOG.rst
@@ -1,0 +1,8 @@
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Changelog for package base2d_kinematics_msgs
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+0.1.0 (2023-08-29)
+------------------
+* Initial kinematics_2d_msgs and kinematics_2d packages
+* Contributors: David V. Lu
