+++ base2d_kinematics_msgs/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kinematics_2d_msgs</name>
+  <name>base2d_kinematics_msgs</name>
   <version>0.1.0</version>
   <description>Interfaces for 2.5D kinematics</description>
   <maintainer email="davidvlu@gmail.com">David V. Lu!!</maintainer>
