--- kinematics_2d_msgs/CHANGELOG.rst
@@ -1,8 +1,0 @@
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Changelog for package kinematics_2d_msgs
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-0.1.0 (2023-08-29)
-------------------
-* Initial kinematics_2d_msgs and kinematics_2d packages
-* Contributors: David V. Lu
